api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [{ name = "Daniel Steman", email = "daniel-steman@live.nl" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.120.2,<0.121.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "PyJWT (>=2.8.0,<3.0.0)",
    "cryptography (>=42.0.0,<43.0.0)",
    "pydantic-settings (>=2.3,<3.0)",
    "psycopg[binary] (>=3.1.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
